+++ Filelist
@@ -111,6 +111,7 @@ SRC_ALL =	\
 		src/spell.c \
 		src/spell.h \
 		src/spellfile.c \
+		src/spellsuggest.c \
 		src/structs.h \
 		src/syntax.c \
 		src/tag.c \
@@ -255,6 +256,7 @@ SRC_ALL =	\
 		src/proto/sound.pro \
 		src/proto/spell.pro \
 		src/proto/spellfile.pro \
+		src/proto/spellsuggest.pro \
 		src/proto/syntax.pro \
 		src/proto/tag.pro \
 		src/proto/term.pro \
